refactor: add addressVersion to network object
