+++ packages/network/src/constants.ts
@@ -34,6 +34,23 @@ export enum TransactionVersion {
   Testnet = 0x80,
 }
 
+/**
+ * Address versions for identifying address types in an encoded Stacks address.
+ * The address version is a single byte, indicating the address type.
+ * Every Stacks address starts with `S` followed by a single character indicating the address version.
+ * The second character is the c32-encoded AddressVersion byte.
+ */
+export enum AddressVersion {
+  /** `P` — A single-sig address for mainnet (starting with `SP`) */
+  MainnetSingleSig = 22,
+  /** `M` — A multi-sig address for mainnet (starting with `SM`) */
+  MainnetMultiSig = 20,
+  /** `T` — A single-sig address for testnet (starting with `ST`) */
+  TestnetSingleSig = 26,
+  /** `N` — A multi-sig address for testnet (starting with `SN`) */
+  TestnetMultiSig = 21,
+}
+
 export const DEFAULT_TRANSACTION_VERSION = TransactionVersion.Mainnet;
 
 /** @ignore */
